feat: Add GroupsAccumulator and GroupsAccumulatorFlatAdapter, extending API
This is based on upstream GroupsAccumulator and
GroupsAccumulatorAdapter, but extends the API so that existing hash
aggregation works with it.  We basically don't really use the upstream
interface (at this time).

We still use basic Accumulator for types that do not implement
GroupsAccumulator, and hash aggregation code handles this poorly.